requirements.txt.txt
tensorflow==2.20.0
streamlit==1.52.2
streamlit-drawable-canvas==0.9.3
numpy
pillow
